+++ src/main/java/com/szs/szsyoungjunkim/refund/application/service/dto/RefundCreateCommand.java
@@ -2,12 +2,12 @@ package com.szs.szsyoungjunkim.refund.application.service.dto;
 
 public record RefundCreateCommand(
         Integer taxYear,
-        Integer totalIncomeTax,
+        Long totalIncomeTax,
         String taxCredit,
         String userId
 ) {
 
-    public static RefundCreateCommand of(Integer taxYear, Integer totalIncomeTax, String taxCredit, String userId) {
+    public static RefundCreateCommand of(Integer taxYear, Long totalIncomeTax, String taxCredit, String userId) {
         return new RefundCreateCommand(taxYear, totalIncomeTax, taxCredit, userId);
     }
 }
