+++ src/main/java/com/szs/szsyoungjunkim/refund/infrastructure/entity/RefundEntity.java
@@ -20,11 +20,11 @@ public class RefundEntity extends BaseTimeEntity {
     private String userId;
 
     private Integer taxYear;
-    private Integer totalIncome;            //종합 소득 금액
+    private Long totalIncome;            //종합 소득 금액
     private Double totalTaxCredit;          //세액 공제
 
     @Builder
-    private RefundEntity(Long id, String userId, Integer taxYear, Integer totalIncome, Double totalTaxCredit) {
+    private RefundEntity(Long id, String userId, Integer taxYear, Long totalIncome, Double totalTaxCredit) {
         this.userId = userId;
         this.taxYear = taxYear;
         this.totalIncome = totalIncome;
